requirements.txt
requests
PyYAML
pyodbc